refactor: OPTIONS 지역 상수로 변경, 시간대 배열 상수명 변경
